Update Data100.md
Prerequisites
While we are working to make this class widely accessible we currently require the following (or equivalent) prerequisites:

Foundations of Data Science: Data 8 covers much of the material in Data 100 but at an introductory level. Data8 provides basic exposure to python programming and working with tabular data as well as visualization, statistics, and machine learning.

Computing: The Structure and Interpretation of Computer Programs CS 61A or Computational Structures in Data Science CS 88. These courses provide additional background in python programming (e.g., for loops, lambdas, debugging, and complexity) that will enable Data 100 to focus more on the concepts in Data Science and less on the details of programming in python.

Math: Linear Algebra (Math 54, EE 16A, or Stat 89A): We will need some basic concepts like linear operators and derivatives to enable statistical inference and derive new prediction algorithms. This may be satisfied concurrently to Data 100.